+++ exportplan/tests/test_views.py
@@ -35,8 +35,8 @@ def export_plan_data(company):
 
 
 @pytest.fixture
-def export_plan():
-    export_plan = factories.CompanyExportPlanFactory.create()
+def export_plan(authed_supplier):
+    export_plan = factories.CompanyExportPlanFactory.create(sso_id=authed_supplier.sso_id)
     factories.CompanyObjectivesFactory.create(companyexportplan=export_plan)
     factories.RouteToMarketsFactory.create(companyexportplan=export_plan)
     factories.TargetMarketDocumentsFactory.create(companyexportplan=export_plan)
@@ -47,11 +47,7 @@ def export_plan():
 
 
 @pytest.mark.django_db
-def test_export_plan_retrieve(authed_client, authed_supplier, export_plan):
-
-    authed_supplier.sso_id = export_plan.sso_id
-    authed_supplier.company = export_plan.company
-    authed_supplier.save()
+def test_export_plan_retrieve(authed_client, export_plan):
     url = reverse('export-plan-detail-update', kwargs={'pk': export_plan.pk})
     response = authed_client.get(url)
 
@@ -135,10 +131,22 @@ def test_export_plan_retrieve(authed_client, authed_supplier, export_plan):
 
 
 @pytest.mark.django_db
-def test_export_plan_update(authed_client, authed_supplier, export_plan):
-    authed_supplier.sso_id = export_plan.sso_id
-    authed_supplier.company = export_plan.company
-    authed_supplier.save()
+def test_export_plan_retrieve_403(authed_client, export_plan):
+    non_owner_exportplan = factories.CompanyExportPlanFactory.create(sso_id=export_plan.sso_id + 1)
+    url = reverse('export-plan-detail-update', kwargs={'pk': non_owner_exportplan.pk})
+    response = authed_client.get(url)
+    assert response.status_code == 403
+
+
+@pytest.mark.django_db
+def test_export_plan_retrieve_not_found(authed_client, export_plan):
+    url = reverse('export-plan-detail-update', kwargs={'pk': 10000})
+    response = authed_client.get(url)
+    assert response.status_code == 404
+
+
+@pytest.mark.django_db
+def test_export_plan_update(authed_client, export_plan):
     url = reverse('export-plan-detail-update', kwargs={'pk': export_plan.pk})
 
     data = {'export_commodity_codes': [{'commodity_name': 'vodka', 'commodity_code': '104.2402.123'}]}
@@ -148,6 +156,16 @@ def test_export_plan_update(authed_client, authed_supplier, export_plan):
     assert response.status_code == http.client.OK
     export_plan.refresh_from_db()
     assert export_plan.export_commodity_codes == data['export_commodity_codes']
+
+
+@pytest.mark.django_db
+def test_export_plan_update_403(authed_client, export_plan):
+    non_owner_exportplan = factories.CompanyExportPlanFactory.create(sso_id=export_plan.sso_id + 1)
+    url = reverse('export-plan-detail-update', kwargs={'pk': non_owner_exportplan.pk})
+
+    data = {'export_commodity_codes': [{'commodity_name': 'vodka', 'commodity_code': '104.2402.123'}]}
+    response = authed_client.patch(url, data, format='json')
+    assert response.status_code == 403
 
 
 @pytest.mark.django_db
@@ -191,12 +209,7 @@ def test_export_plan_create(authed_client, authed_supplier):
 
 @pytest.mark.django_db
 def test_export_plan_update_json_new_to_partial(authed_client, authed_supplier):
-    export_plan = factories.CompanyExportPlanFactory.create(about_your_business={})
-
-    authed_supplier.sso_id = export_plan.sso_id
-    authed_supplier.company = export_plan.company
-    authed_supplier.save()
-
+    export_plan = factories.CompanyExportPlanFactory.create(about_your_business={}, sso_id=authed_supplier.sso_id)
     url = reverse('export-plan-detail-update', kwargs={'pk': export_plan.pk})
     response = authed_client.get(url)
 
@@ -223,12 +236,9 @@ def test_export_plan_update_json_new_to_partial(authed_client, authed_supplier):
 
 @pytest.mark.django_db
 def test_export_plan_update_json_new_to_partial_inner_dict(authed_client, authed_supplier):
-    export_plan = factories.CompanyExportPlanFactory.create(ui_progress={'section-a': {'opt-a': 'A', 'opt-b': 'B'}})
-
-    authed_supplier.sso_id = export_plan.sso_id
-    authed_supplier.company = export_plan.company
-    authed_supplier.save()
-
+    export_plan = factories.CompanyExportPlanFactory.create(
+        ui_progress={'section-a': {'opt-a': 'A', 'opt-b': 'B'}}, sso_id=authed_supplier.sso_id
+    )
     url = reverse('export-plan-detail-update', kwargs={'pk': export_plan.pk})
     response = authed_client.get(url)
     assert response.status_code == 200
@@ -258,11 +268,7 @@ def test_export_plan_update_json_new_to_partial_inner_dict(authed_client, authed
 
 @pytest.mark.django_db
 def test_export_plan_update_non_json_new_to_partial(authed_client, authed_supplier):
-    export_plan = factories.CompanyExportPlanFactory.create(adaptation_target_market={})
-
-    authed_supplier.sso_id = export_plan.sso_id
-    authed_supplier.company = export_plan.company
-    authed_supplier.save()
+    export_plan = factories.CompanyExportPlanFactory.create(adaptation_target_market={}, sso_id=authed_supplier.sso_id)
 
     url = reverse('export-plan-detail-update', kwargs={'pk': export_plan.pk})
     response = authed_client.get(url)
@@ -317,10 +323,7 @@ def test_export_plan_update_non_json_new_to_partial(authed_client, authed_suppli
     ],
 )
 @pytest.mark.django_db
-def test_export_plan_model_create(model_class, property_name, create_data, authed_client, authed_supplier, export_plan):
-    authed_supplier.sso_id = export_plan.sso_id
-    authed_supplier.company = export_plan.company
-    authed_supplier.save()
+def test_export_plan_model_create(model_class, property_name, create_data, authed_client, export_plan):
     url = reverse('export-plan-model-object-list-create')
 
     create_data['companyexportplan'] = export_plan.id
@@ -352,10 +355,7 @@ def test_export_plan_model_create(model_class, property_name, create_data, authe
     ],
 )
 @pytest.mark.django_db
-def test_export_plan_model_retrieve(model_class, property_name, authed_client, authed_supplier, export_plan):
-    authed_supplier.sso_id = export_plan.sso_id
-    authed_supplier.company = export_plan.company
-    authed_supplier.save()
+def test_export_plan_model_retrieve(model_class, property_name, authed_client, export_plan):
 
     model_object = getattr(export_plan, property_name).all()[0]
 
@@ -385,10 +385,8 @@ def test_export_plan_model_retrieve(model_class, property_name, authed_client, a
     ],
 )
 @pytest.mark.django_db
-def test_export_plan_model_update(model_class, property_name, data_update, authed_client, authed_supplier, export_plan):
-    authed_supplier.sso_id = export_plan.sso_id
-    authed_supplier.company = export_plan.company
-    authed_supplier.save()
+def test_export_plan_model_update(model_class, property_name, data_update, authed_client, export_plan):
+
     attribute_updated = list(data_update.keys())[0]
     data_update['model_name'] = model_class.__name__
     model_object = getattr(export_plan, property_name).all()[0]
@@ -413,10 +411,8 @@ def test_export_plan_model_update(model_class, property_name, data_update, authe
     ],
 )
 @pytest.mark.django_db
-def test_export_plan_model_delete(model_class, property_name, authed_client, authed_supplier, export_plan):
-    authed_supplier.sso_id = export_plan.sso_id
-    authed_supplier.company = export_plan.company
-    authed_supplier.save()
+def test_export_plan_model_delete(model_class, property_name, authed_client, export_plan):
+
     model_object = getattr(export_plan, property_name).all()[0]
 
     # Upper model name is deliberate to test that it's not case sensitive
@@ -429,10 +425,7 @@ def test_export_plan_model_delete(model_class, property_name, authed_client, aut
 
 
 @pytest.mark.django_db
-def test_export_plan_pdf_upload(authed_client, authed_supplier, export_plan):
-    authed_supplier.sso_id = export_plan.sso_id
-    authed_supplier.company = export_plan.company
-    authed_supplier.save()
+def test_export_plan_pdf_upload(authed_client, export_plan):
     mock_file = mock.Mock(spec=File)
     url = reverse('export-plan-pdf-upload')
 
@@ -443,3 +436,14 @@ def test_export_plan_pdf_upload(authed_client, authed_supplier, export_plan):
     export_plan_upload = models.ExportplanDownloads.objects.last()
     assert export_plan_upload.companyexportplan.id == export_plan.id
     assert export_plan_upload.pdf_file is not None
+
+
+@pytest.mark.django_db
+def test_export_plan_pdf_upload_403(authed_client, export_plan):
+    mock_file = mock.Mock(spec=File)
+    url = reverse('export-plan-pdf-upload')
+    non_owner_exportplan = factories.CompanyExportPlanFactory.create(sso_id=export_plan.sso_id + 1)
+    data = {'companyexportplan': non_owner_exportplan.id, 'pdf_file': mock_file}
+
+    response = authed_client.post(url, data)
+    assert response.status_code == 403